+++ splashcore/animusmagic/messages.go
@@ -6,6 +6,57 @@ import (
 	"github.com/anti-raid/splashtail/splashcore/types"
 )
 
+/*
+   pub struct AmCheckCommandOptionsFlags: u8 {
+       /// Whether or not to ignore the cache
+       const IGNORE_CACHE = 1 << 0;
+       /// Whether or not to cache the result at all
+       const CACHE_RESULT = 1 << 1;
+       /// Whether or not to ignore the fact that the module is disabled in the guild
+       const IGNORE_MODULE_DISABLED = 1 << 2;
+       /// Whether or not to ignore the fact that the command is disabled in the guild
+       const IGNORE_COMMAND_DISABLED = 1 << 3;
+       /// Skip custom resolved kittycat permission fit 'checks' (AKA does the user have the actual permissions ofthe custom resolved permissions)
+       const SKIP_CUSTOM_RESOLVED_FIT_CHECKS = 1 << 4;
+   }
+
+   /// Flags of type AmCheckCommandOptionsFlags
+   #[serde(default)]
+   pub flags: u8,
+
+   /// What custom resolved permissions to use for the user. Note that ensure_user_has_custom_resolved must be true to ensure that the user has all the permissions in the custom_resolved_kittycat_perms
+   ///
+   /// API needs this for limiting the permissions of a user, allows setting custom resolved perms
+   #[serde(default)]
+   pub custom_resolved_kittycat_perms: Option<Vec<String>>,
+
+   /// Custom permission checks to use
+   #[serde(default)]
+   pub custom_command_configuration: Option<Box<silverpelt::GuildCommandConfiguration>>,
+
+   /// Custom permission checks to use
+   #[serde(default)]
+   pub custom_module_configuration: Option<Box<silverpelt::GuildModuleConfiguration>>,
+
+*/
+
+type AmCheckCommandOptionsFlag uint8
+
+const (
+	AmCheckCommandOptionsFlagIgnoreCache                 AmCheckCommandOptionsFlag = 1 << 0
+	AmCheckCommandOptionsFlagCacheResult                 AmCheckCommandOptionsFlag = 1 << 1
+	AmCheckCommandOptionsFlagIgnoreModuleDisabled        AmCheckCommandOptionsFlag = 1 << 2
+	AmCheckCommandOptionsFlagIgnoreCommandDisabled       AmCheckCommandOptionsFlag = 1 << 3
+	AmCheckCommandOptionsFlagSkipCustomResolvedFitChecks AmCheckCommandOptionsFlag = 1 << 4
+)
+
+type AmCheckCommandOptions struct {
+	Flags                       AmCheckCommandOptionsFlag             `json:"flags"`
+	CustomResolvedKittycatPerms *[]string                             `json:"custom_resolved_kittycat_perms,omitempty"`
+	CustomCommandConfiguration  *silverpelt.GuildCommandConfiguration `json:"custom_command_configuration,omitempty"`
+	CustomModuleConfiguration   *silverpelt.GuildModuleConfiguration  `json:"custom_module_configuration,omitempty"`
+}
+
 type ClusterModules = []silverpelt.CanonicalModule
 
 type BotAnimusMessage struct {
@@ -18,10 +69,10 @@ type BotAnimusMessage struct {
 		UserID  string `json:"user_id"`
 	} `json:"BaseGuildUserInfo,omitempty"`
 	CheckCommandPermission *struct {
-		GuildID             string                         `json:"guild_id"`
-		UserID              string                         `json:"user_id"`
-		Command             string                         `json:"command"`
-		CheckCommandOptions silverpelt.CheckCommandOptions `json:"opts"`
+		GuildID             string                `json:"guild_id"`
+		UserID              string                `json:"user_id"`
+		Command             string                `json:"command"`
+		CheckCommandOptions AmCheckCommandOptions `json:"opts"`
 	} `json:"CheckCommandPermission,omitempty"`
 	ExecutePerModuleFunction *struct {
 		Module  string         `json:"module"`