+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Bot
 
 - The bot now runs permission checks for reset toggle on both commands and modules based on the default state of the command/module. In addition, commands/modules must be toggleable to be able to reset toggles.
+- Refactored permission checking a bit internally. Users should see no changes beyond potentially improved performance due to using sandwich state more.
+- Added `virtual_command` to `CommandExtendedData` struct. Virtual commands, like virtual modules, are not actually loaded into the bot but can be used for permission checks etc.
+- The ``acl__{module}_defaultperms_check`` command has been added to better handle default permission checks
+- ``custom_module_configuration`` has been added to ``can_run_command``
+- The `AmCheckCommandOptions` struct used for permission checks has been changed significantly. It now uses a ``flags`` bitfield (`u8`) instead of 5-6 bool fields to save memory. **These changes have also been ported to the webserver as well**
+- Added ``Module.parse`` to allow performing some checks on the modules before starting up the bot in an invalid state
+- ``ensure_custom_kittycat_perms`` has been replaced with ``SKIP_CUSTOM_RESOLVED_FIT_CHECKS``. The new API is opt-in hence ensuring that all permission limits are checked by default unless explicitly overriden instead of the less secure opt-out approach of `ensure_custom_kittycat_perms`.
 
 ## Website